+++ CHANGES.md
@@ -10,9 +10,10 @@
 ## 2021.5.8
 ### New Features
 - When creating a DNS record to point a hostname at a tunnel, you can now use --overwrite-dns to overwrite any existing
-  DNS records with that hostname. This works when using the CLI to provision DNS, and when starting an adhoc tunnel, e.g.
-  - `cloudflared route --overwrite-dns dns foo foo.example.com`
-  - `cloudflared tunnel --overwrite-dns --name foo --hostname foo.example.com`
+  DNS records with that hostname. This works both when using the CLI to provision DNS, as well as when starting an adhoc
+  named tunnel, e.g.:
+  - `cloudflared tunnel route dns --overwrite-dns foo-tunnel foo.example.com`
+  - `cloudflared tunnel --overwrite-dns --name foo-tunnel --hostname foo.example.com`
 
 ## 2021.5.7
 ### New Features